all prime factors of a number
This is a fast method to find number of prime factors and print them. Actually, this is a modification of  "Sieve of Eratosthenes".

This program contain 2 functions. 
First function, calculate number of prime factors of a number and simultaneously finding if a number is prime and also the largest prime factor of the number.
Second function, print all prime factors of a number.

Due to memory limit, variable "siz" can hold maximum value of 10000000.